backend/FastAPI-project-template/requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.63.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.0.0
h11==0.12.0
importlib-metadata==4.0.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
PyMySQL==1.0.2
python-dotenv==0.17.0
python-jose==3.2.0
PyYAML==5.4.1
rsa==4.7.2
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.15.0
SQLAlchemy==1.4.9
alembic==1.5.8
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
zipp==3.4.1
Faker==7.0.0
dnspython==2.1.0
docstring-parser==0.7.3
email-validator==1.1.2
text-unidecode==1.3
python-editor==1.0.4
python-multipart==0.0.5
requests==2.25.1
python-dateutil==2.8.1